refactor(parser): extract token-to-node pipeline
